+++ pkg/scalers/selenium_grid_scaler.go
@@ -35,6 +35,7 @@ type seleniumGridScalerMetadata struct {
 	browserVersion      string
 	unsafeSsl           bool
 	scalerIndex         int
+	platformName        string
 }
 
 type seleniumResponse struct {
@@ -65,10 +66,12 @@ type seleniumSession struct {
 type capability struct {
 	BrowserName    string `json:"browserName"`
 	BrowserVersion string `json:"browserVersion"`
+	PlatformName   string `json:"platformName"`
 }
 
 const (
 	DefaultBrowserVersion string = "latest"
+	DefaultPlatformName   string = "linux"
 )
 
 func NewSeleniumGridScaler(config *ScalerConfig) (Scaler, error) {
@@ -143,6 +146,12 @@ func parseSeleniumGridScalerMetadata(config *ScalerConfig) (*seleniumGridScalerM
 		meta.unsafeSsl = parsedVal
 	}
 
+	if val, ok := config.TriggerMetadata["platformName"]; ok && val != "" {
+		meta.platformName = val
+	} else {
+		meta.platformName = DefaultPlatformName
+	}
+
 	meta.scalerIndex = config.ScalerIndex
 	return &meta, nil
 }
@@ -206,14 +215,14 @@ func (s *seleniumGridScaler) getSessionsCount(ctx context.Context, logger logr.L
 	if err != nil {
 		return -1, err
 	}
-	v, err := getCountFromSeleniumResponse(b, s.metadata.browserName, s.metadata.browserVersion, s.metadata.sessionBrowserName, logger)
+	v, err := getCountFromSeleniumResponse(b, s.metadata.browserName, s.metadata.browserVersion, s.metadata.sessionBrowserName, s.metadata.platformName, logger)
 	if err != nil {
 		return -1, err
 	}
 	return v, nil
 }
 
-func getCountFromSeleniumResponse(b []byte, browserName string, browserVersion string, sessionBrowserName string, logger logr.Logger) (int64, error) {
+func getCountFromSeleniumResponse(b []byte, browserName string, browserVersion string, sessionBrowserName string, platformName string, logger logr.Logger) (int64, error) {
 	var count int64
 	var seleniumResponse = seleniumResponse{}
 
@@ -226,9 +235,10 @@ func getCountFromSeleniumResponse(b []byte, browserName string, browserVersion s
 		var capability = capability{}
 		if err := json.Unmarshal([]byte(sessionQueueRequest), &capability); err == nil {
 			if capability.BrowserName == browserName {
-				if strings.HasPrefix(capability.BrowserVersion, browserVersion) {
+				var platformNameMatches = capability.PlatformName == "" || strings.EqualFold(capability.PlatformName, platformName)
+				if strings.HasPrefix(capability.BrowserVersion, browserVersion) && platformNameMatches {
 					count++
-				} else if browserVersion == DefaultBrowserVersion {
+				} else if browserVersion == DefaultBrowserVersion && platformNameMatches {
 					count++
 				}
 			}
@@ -241,10 +251,11 @@ func getCountFromSeleniumResponse(b []byte, browserName string, browserVersion s
 	for _, session := range sessions {
 		var capability = capability{}
 		if err := json.Unmarshal([]byte(session.Capabilities), &capability); err == nil {
+			var platformNameMatches = capability.PlatformName == "" || strings.EqualFold(capability.PlatformName, platformName)
 			if capability.BrowserName == sessionBrowserName {
-				if strings.HasPrefix(capability.BrowserVersion, browserVersion) {
+				if strings.HasPrefix(capability.BrowserVersion, browserVersion) && platformNameMatches {
 					count++
-				} else if browserVersion == DefaultBrowserVersion {
+				} else if browserVersion == DefaultBrowserVersion && platformNameMatches {
 					count++
 				}
 			}